remove duplicate http-error test
