+++ .changeset/tough-pets-sit.md
@@ -1,0 +1,5 @@
+---
+'@yeliex/fastify-problem-details': patch
+---
+
+Consolidate duplicate http-error tests by keeping Fastify integration coverage in the Fastify package tests.
